Organize all component files into components folder
