+++ src/Components/Favorites/Favorites.css
