+++ src/Components/QuoteCard/QuoteCard.css
