+++ src/Components/Quotes/Quotes.css
