+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from "react-router-dom";
 import './index.css';
-import App from './App/App';
+import App from './Components/App/App';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
